perf(service): reduce the interaction of closeTransactionById with the repositories

+++ src/main/java/com/onspring/onspring_customer/domain/common/service/TransactionArchiveService.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Pageable;
 import java.util.List;
 
 public interface TransactionArchiveService {
-    List<Long> closeTransactionById(List<Long> ids);
+    List<Integer> closeTransactionById(List<Long> ids);
 
     Page<TransactionArchiveDto> findAllTransactionArchive(Pageable pageable);
 }
